Rename the main component file to 'index.tsx'.
